# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Archive API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dashboard/summary": {
      "get": {
        "summary": "Get Dashboard Summary",
        "operationId": "get_dashboard_summary_api_dashboard_summary_get",
        "parameters": [
          {
            "name": "recent_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Recent Limit"
            }
          },
          {
            "name": "pinned_per_category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Pinned Per Category"
            }
          },
          {
            "name": "pinned_category_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Pinned Category Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/tasks": {
      "get": {
        "summary": "Get Dashboard Tasks",
        "operationId": "get_dashboard_tasks_api_dashboard_tasks_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM",
              "title": "Month"
            },
            "description": "YYYY-MM"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Dashboard Task",
        "operationId": "create_dashboard_task_api_dashboard_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardTaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTaskItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/tasks/{task_id}": {
      "get": {
        "summary": "Get Dashboard Task",
        "operationId": "get_dashboard_task_api_dashboard_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTaskItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Dashboard Task",
        "operationId": "update_dashboard_task_api_dashboard_tasks__task_id__patch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardTaskUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTaskItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Dashboard Task",
        "operationId": "delete_dashboard_task_api_dashboard_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/task-settings": {
      "get": {
        "summary": "Get Dashboard Task Settings",
        "operationId": "get_dashboard_task_settings_api_dashboard_task_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTaskSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Dashboard Task Settings",
        "operationId": "update_dashboard_task_settings_api_dashboard_task_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardTaskSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardTaskSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardCategoryCount": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "category",
          "count"
        ],
        "title": "DashboardCategoryCount"
      },
      "DashboardErrorCodeCount": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "count"
        ],
        "title": "DashboardErrorCodeCount"
      },
...
      "DashboardSummaryResponse": {
        "properties": {
          "total_documents": {
            "type": "integer",
            "title": "Total Documents"
          },
          "recent_uploads_7d": {
            "type": "integer",
            "title": "Recent Uploads 7D"
          },
          "needs_review_count": {
            "type": "integer",
            "title": "Needs Review Count"
          },
          "failed_jobs_count": {
            "type": "integer",
            "title": "Failed Jobs Count"
          },
          "retry_scheduled_count": {
            "type": "integer",
            "title": "Retry Scheduled Count"
          },
          "dead_letter_count": {
            "type": "integer",
            "title": "Dead Letter Count"
          },
          "failed_error_codes": {
            "items": {
              "$ref": "#/components/schemas/DashboardErrorCodeCount"
            },
            "type": "array",
            "title": "Failed Error Codes"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/DashboardCategoryCount"
            },
            "type": "array",
            "title": "Categories"
          },
          "pinned_by_category": {
            "items": {
              "$ref": "#/components/schemas/DashboardPinnedCategory"
            },
            "type": "array",
            "title": "Pinned By Category"
          },
          "recent_documents": {
            "items": {
              "$ref": "#/components/schemas/DashboardRecentDocument"
            },
            "type": "array",
            "title": "Recent Documents"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "total_documents",
          "recent_uploads_7d",
          "needs_review_count",
          "failed_jobs_count",
          "retry_scheduled_count",
          "dead_letter_count",
          "generated_at"
        ],
        "title": "DashboardSummaryResponse"
      },
      "DashboardTaskCreateRequest": {
        "properties": {
          "category": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Category"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 1,
            "title": "Title"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "all_day": {
            "type": "boolean",
            "title": "All Day",
            "default": false
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 220
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "category",
          "title",
          "scheduled_at"
        ],
        "title": "DashboardTaskCreateRequest"
      },
      "DashboardTaskItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "all_day": {
            "type": "boolean",
            "title": "All Day",
            "default": false
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "title",
          "scheduled_at"
        ],
        "title": "DashboardTaskItem"
      },
      "DashboardTaskListResponse": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DashboardTaskItem"
            },
            "type": "array",
            "title": "Items"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "month",
          "generated_at"
        ],
        "title": "DashboardTaskListResponse"
      },
...
      "DashboardTaskUpdateRequest": {
        "properties": {
          "category": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Category"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 1,
            "title": "Title"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "all_day": {
            "type": "boolean",
            "title": "All Day",
            "default": false
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 220
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "category",
          "title",
          "scheduled_at"
        ],
        "title": "DashboardTaskUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}